DCWL-1627 Fix failing test, minor touch-ups

+++ app/uk/gov/hmrc/customs/notification/controllers/Results.scala
@@ -22,7 +22,7 @@ import play.api.mvc.Results.{Ok, Status}
 import play.mvc.Http.Status.*
 
 private object Results {
-  def blockedCountOkResponseFrom(count: Int): Result = {
+  def blockedCountOkResponseFrom(count: Long): Result = {
     val countXml =
       <pushNotificationBlockedCount>
         {count}
